improves snapshot serialization and tests
