+++ blockstore/apps/api/v1/serializers/bundles.py
@@ -7,6 +7,7 @@ from rest_framework.relations import SlugRelatedField
 from rest_framework.reverse import reverse
 
 from blockstore.apps.bundles.models import Bundle, BundleVersion, Collection
+from blockstore.apps.bundles.store import SnapshotRepo
 from ... import relations
 
 
@@ -78,6 +79,53 @@ class BundleSerializer(serializers.ModelSerializer):
     )
 
 
+class SnapshotSerializer(serializers.Serializer):
+    """Serializer for a snapshot object."""
+    # pylint: disable=abstract-method
+
+    def _expand_url(self, url):
+        """Ensure that the given URL is an absolute URL"""
+        if not url.startswith('http'):
+            url = self.context['request'].build_absolute_uri(url)
+        return url
+
+    def _serialized_dep(self, dependency):
+        return {
+            "bundle_uuid": dependency.bundle_uuid,
+            "version": dependency.version,
+            "snapshot_digest": dependency.snapshot_digest.hex(),
+        }
+
+    def to_representation(self, instance):
+        snapshot_repo = SnapshotRepo()
+        info = {
+            'hash_digest': instance.hash_digest.hex(),
+            'created_at': instance.created_at,
+        }
+
+        info['files'] = {
+            path: {
+                "url": self._expand_url(snapshot_repo.url(instance, path)),
+                "size": file_info.size,
+                "hash_digest": file_info.hash_digest.hex(),
+            }
+            for path, file_info in instance.files.items()
+        }
+
+        info['links'] = {
+            link.name: {
+                "direct": self._serialized_dep(link.direct_dependency),
+                "indirect": [
+                    self._serialized_dep(dep)
+                    for dep in link.indirect_dependencies
+                ]
+            }
+            for link in instance.links
+        }
+
+        return info
+
+
 class BundleVersionSerializer(serializers.ModelSerializer):
     """Serializer for the BundleVersion model."""
 
@@ -134,7 +182,7 @@ class BundleVersionWithFileDataSerializer(BundleVersionSerializer):
         def to_representation(self, value):
             """Snapshot JSON serialization."""
             view_kwargs = self.context['view'].kwargs
-            info = value.serialize(context=self.context)
+            info = SnapshotSerializer(value, context=self.context).data
             for file_name, file_info in info['files'].items():
                 file_info['url'] = self._expand_url('{}?file={}'.format(
                     reverse('api:v1:bundleversion-redirect', kwargs=view_kwargs),
